backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Task Queue
redis==5.0.1
rq==1.16.0
rq-scheduler==0.13.1
APScheduler==3.10.4

# Authentication
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# HTTP client & RSS
httpx==0.26.0
feedparser==6.0.11

# OpenAI
openai==1.10.0

# Validation & serialization
pydantic[email]==2.5.3
pydantic-settings==2.1.0
pyyaml==6.0.1
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

